requirements.txt
openpyxl>=3.1.2
reportlab>=4.0.7
pypdf>=3.17.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
